added placard modal to show art
