+++ src/components/artwork/IndexArt.js
@@ -18,7 +18,6 @@ class IndexArt extends Component {
       .then(response => this.setState({
         artworks: response.data.artworks
       }))
-      // .then(console.log('user logged in is:', this.props.user))
       .catch(error => {
         this.setState({
           artworks: null
